+++ src/ncFile.ts
@@ -79,4 +79,23 @@ export default class NCFile {
     public async addTag(tagName: string): Promise<void> {
         return this.client.addTagToFile(await this.getId(), tagName);
     }
+
+    /**
+     * add comment to file
+     * @param comment the comment
+     */
+    public async addComment(comment: string): Promise<void> {
+        return this.client.addCommentToFile(await this.getId(), comment);
+    }
+
+    /**
+     * get list of comments of file
+     * @param top number of comments to return
+     * @param skip the offset
+     * @returns array of comment strings
+     * @throws Exception
+     */
+    public async getComments(top?: number, skip?: number): Promise<string[]> {
+        return this.client.getFileComments(await this.getId(), top, skip);
+    }
 }